perf: Reduce imports by moving bookings zod out of utils
