+++ apps/api/v1/lib/validations/booking.ts
@@ -8,7 +8,8 @@ import {
   _TeamModel,
   _UserModel,
 } from "@calcom/prisma/zod";
-import { extendedBookingCreateBody, iso8601 } from "@calcom/prisma/zod-utils";
+import { iso8601 } from "@calcom/prisma/zod-utils";
+import { extendedBookingCreateBody } from "@calcom/prisma/zod/custom/booking";
 
 import { schemaQueryUserId } from "./shared/queryUserId";
