+++ packages/features/bookings/lib/getBookingDataSchema.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 
-import { extendedBookingCreateBody } from "@calcom/prisma/zod-utils";
+import { extendedBookingCreateBody } from "@calcom/prisma/zod/custom/booking";
 
 import type { getBookingFieldsWithSystemFields } from "./getBookingFields";
 import getBookingResponsesSchema from "./getBookingResponsesSchema";
