+++ packages/features/bookings/lib/handleNewBooking/getBookingData.ts
@@ -5,7 +5,7 @@ import type z from "zod";
 import dayjs from "@calcom/dayjs";
 import { getCalEventResponses } from "@calcom/features/bookings/lib/getCalEventResponses";
 import { ErrorCode } from "@calcom/lib/errorCodes";
-import { bookingCreateSchemaLegacyPropsForApi } from "@calcom/prisma/zod-utils";
+import { bookingCreateSchemaLegacyPropsForApi } from "@calcom/prisma/zod/custom/booking";
 
 import type { TgetBookingDataSchema } from "../getBookingDataSchema";
 import { handleCustomInputs } from "./handleCustomInputs";
