+++ packages/platform/atoms/hooks/bookings/useHandleBookEvent.ts
@@ -7,7 +7,7 @@ import { mapBookingToMutationInput, mapRecurringBookingToMutationInput } from "@
 import type { BookerEvent } from "@calcom/features/bookings/types";
 import { useLocale } from "@calcom/lib/hooks/useLocale";
 import type { RoutingFormSearchParams } from "@calcom/platform-types";
-import type { BookingCreateBody } from "@calcom/prisma/zod-utils";
+import type { BookingCreateBody } from "@calcom/prisma/zod/custom/booking";
 import { showToast } from "@calcom/ui";
 
 import type { UseCreateBookingInput } from "./useCreateBooking";
